# auth-api/api/openapi.yaml
openapi: 3.0.1
info:
  title: OAuth Server 
  version: 1.1.0
servers:
- url: /
tags:
- name: Users
  description: Operations related with the manipulation of users raw data
paths:
  /api/v1/user/{user_id}:
    get:
      tags:
      - Users
      summary: Get a User information
      description: Retrieve information of a specific User
      operationId: getUser
      parameters:
      - name: user_id
        in: path
        description: ID of a user
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/user_id'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        "400":
          description: The provided GBFS was not found
          content:
            text/html:
              schema:
                $ref: '#/components/schemas/errorMessage'
              example: Inexistent User Id
      x-swagger-router-controller: Users
    post:
      tags:
      - Users
      summary: Posts a User information
      description: Posts new user information for saving it
      operationId: postUser
      parameters:
      - name: user_id
        in: path
        description: ID of a user
        required: true
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/user_id'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                $ref: '#/components/schemas/okMessage'
        "400":
          description: The provided id for user is incorrect
          content:
            text/html:
              schema:
                $ref: '#/components/schemas/errorMessage'
              examples:
                invalidAttributes:
                  value: Not a valid ID.
      x-swagger-router-controller: Users
components:
  schemas:
    user_id:
      type: string
    user:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/user_id'
        full_name:
          maxLength: 256
          type: string
          description: Full name os the user
        email:
          maxLength: 512
          type: string
          description: Email of the user
      example:
        full_name: full_name
        _id: id
        email: email
    errorMessage:
      maxLength: 256
      type: string
      description: A text message describing an error
    okMessage:
      type: string
      description: A message that describes an OK response
  parameters:
    user_id:
      name: user_id
      in: path
      description: ID of a user
      required: true
      style: simple
      explode: false
      schema:
        $ref: '#/components/schemas/user_id'